LayoutTests/media/modern-media-controls/placard-support/placard-support-error-expected.txt

Testing the PlacardSupport behavior with an error.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mediaController.controls.placard is mediaController.controls.invalidPlacard
PASS mediaController.controls.children.includes(mediaController.controls.invalidPlacard) is true

PASS successfullyParsed is true

TEST COMPLETE
